Moved interaction-related files into Interactables folder
