--- Assets/_app/Scripts/EventOnlyInteractable.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 257d066a5fe401a42a62825224a5150c
