--- Assets/_app/Scripts/Interactable.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 97c3723599f1b344b87062893e595ab4
