+++ Assets/_app/Scripts/Interactables/EventOnlyInteractable.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 7bb7ec474cd645a41a99cf4493302cdd
