+++ Assets/_app/Scripts/Interactables/Interactable.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 2f340db3d4916764385a49f262fbaa63
