+++ Assets/_app/Scripts/Interactables/InteractionEvent.cs
@@ -1,7 +1,7 @@
 using UnityEngine;
 using UnityEngine.Events;
 
-namespace _app.Scripts
+namespace _app.Scripts.Interactables
 {
     public class InteractionEvent : MonoBehaviour
     {
