+++ Assets/_app/Scripts/Interactables/InteractionEvent.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: d86babba146457641b636c44f122e8bf
