--- Assets/_app/Scripts/InteractionEvent.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: ba3525e0382ae9143b70d50ca587534d
